Rename sdecdata class to SimulationPacketData and shift it outside of sdec_plot.py
